Adding tabs to reduce the amount of scrolling.
I used the all css method of adding tabs to dashboard that was provided by the late great Niket Nilay from this answer response: https://community.splunk.com/t5/Dashboards-Visualizations/How-can-we-create-Tabs-using-Link-List-in-Splunk-with-only-CSS/td-p/485176.  I find it easier to navigate the dashboard as the tabs reduces the amount of scrolling necessary to get to the section I want to investigate.